+++ CLI11.hpp.in
@@ -62,6 +62,8 @@ namespace {namespace} {{
 
 {app_hpp}
 
+{app_inl_hpp}
+
 {config_hpp}
 
 {formatter_hpp}
